float->int

+++ tcp-generator/calc_parameters.py
@@ -6,7 +6,7 @@ def calc_parameters(tput):
     return {'batch_size':b,'wait_time':wt}
 
 def calc_batch_size(tput):
-    return math.floor(tput / 10000.0) + 1
+    return int(math.floor(tput / 10000.0) + 1)
 
 def calc_wait_time(tput, b):
     return b * (1000.0 / tput)
